fix(cli): list Karta definitions that name no workload type
A Karta whose rootComponent omits kind is valid against the CRD, whose
rootComponent.required is [name] alone, but it cannot be indexed by GVK.
The resolver dropped it silently, so an operator saw their definition in
"kubectl get kartas" and nothing at all in "karta definitions". Such
definitions are now listed after the ones that map, addressable by name,
and never matched by kind.

Also simplifies the resolver following review:

- Collision carries Names instead of Winner and Shadowed. List and
  ByRootKind return every claimant, so nothing is discarded at this
  layer and the warning no longer claims to ignore a definition it
  listed on the previous line.
- Drops deepCopy from the accessors, and the defensive clones from List
  and Collisions. A resolver is built per command invocation and no
  caller mutates what it is handed.
- index collects one source's claims into a single local map rather
  than tracking replacement with a second map.
- Folds cluster.go into load.go, keeping the pure lookup in resolver.go
  free of any client-go import.
- Renames ByKind to ByRootKind, which matches the Kind segment of a GVK
  that callers then narrow by version and group.
